requirements.txt
certifi==2017.7.27.1
future==0.16.0
PyGreSQL==5.0.4
python-telegram-bot==8.0
PyYAML==3.12
